Reduce BV calculator release-maintenance debt
Constraint: BV parity must remain source-truth guarded while shrinking the validator core.
Rejected: Rewriting calculateUnitBV formula flow | too broad for a maintenance debt slice.
Confidence: high
Scope-risk: moderate
Directive: Keep future BV calculator splits parity-first; do not move defensive/offensive correction logic without validate:bv evidence.
Tested: npm.cmd run validate:bv; npm.cmd run typecheck -- --pretty false; npm.cmd run maintain:scan -- --scope=scripts/validate-bv-calculator.ts,scripts/validate-bv-unit-derived.ts,scripts/validate-bv-unit-overrides.ts,scripts/validate-bv-movement-derived.ts --json; npm.cmd run maintain:scan:gate; node scripts/qc/validate-maintenance-warning-ledger.mjs; npm.cmd run qc:validate; npm.cmd run qc:app-completion -- --json; git diff --check
Not-tested: Full verify:qc and Electron build were not rerun for this maintenance-only script extraction.
