+++ scripts/validate-bv-unit-overrides.ts
@@ -1,0 +1,66 @@
+export const KNOWN_RISC_OVERRIDE_KIT_UNITS = new Set([
+  'battleaxe-bkx-risc',
+  'emperor-emp-6x',
+  'mad-cat-mk-iv-pr-risc',
+  'malice-mal-y-sh-risc',
+]);
+
+export const KNOWN_HD_GYRO_UNITS = new Set([
+  'albatross-alb-5u',
+  'albatross-alb-5w',
+  'albatross-alb-5w-dantalion',
+  'atlas-as8-k',
+  'atlas-as8-ke',
+  'atlas-ii-as7-d-h-devlin',
+  'battlemaster-blr-10s',
+  'battlemaster-blr-10s2',
+  'battlemaster-blr-k4',
+  'cougar-x-3',
+  'deva-c-dva-o-achilleus',
+  'deva-c-dva-o-invictus',
+  'deva-c-dva-oa-dominus',
+  'deva-c-dva-ob-infernus',
+  'deva-c-dva-oc-comminus',
+  'deva-c-dva-od-luminos',
+  'deva-c-dva-oe-eminus',
+  'deva-c-dva-os-caelestis',
+  'deva-c-dva-ou-exanimus',
+  'grand-titan-t-it-n13m',
+  'griffin-grf-4n',
+  'griffin-grf-5k',
+  'hunchback-hbk-5ss',
+  'jade-hawk-jhk-04',
+  'king-crab-kgc-008',
+  'king-crab-kgc-008b',
+  'king-crab-kgc-009',
+  'king-crab-kgc-009c',
+  'marauder-mad-9w',
+  'marauder-mad-9w2',
+  'ostroc-osr-4k',
+  'ostsol-otl-9r',
+  'patriot-pkm-2c',
+  'patriot-pkm-2d',
+  'patriot-pkm-2e',
+  'phoenix-hawk-pxh-4w',
+  'phoenix-hawk-pxh-99',
+  'scourge-scg-wx1',
+  'sentry-snt-w5',
+  'tai-sho-tsh-8s',
+  'templar-iii-tlr2-j-arthur',
+  'templar-iii-tlr2-o',
+  'templar-iii-tlr2-oa',
+  'templar-iii-tlr2-ob',
+  'templar-iii-tlr2-oc',
+  'templar-iii-tlr2-od',
+  'thunderbolt-iic-2',
+  'thunderbolt-tdr-11s',
+  'thunderbolt-tdr-17s',
+  'thunderbolt-tdr-7s',
+  'vanquisher-vqr-5v',
+  'vanquisher-vqr-7v',
+  'victor-vtr-12d',
+  'warhammer-whd-10ct',
+  'warhammer-whm-x7-the-lich',
+  'white-flame-whf-3c',
+  'xanthos-xnt-4o',
+]);
